Enhancements for next preview (#102)
* add configuration exporters.

Also change signature for crescendo command constructor to include the original command

* bump version

set default target for export-crescendocommand

* Point to web version of schema rather than the one on disk.

This means that the schema is not needed on disk to develop and you don't need to track relative path to schema.
Also add test for new behavior around ConfirmImpact.

* Add test for ConfirmImpact

Update schema with more descriptions. Add support for ConfirmImpact in generator.