(AB-1952368) Add issue templates (#158)
Prior to this change, the repository had no issue templates. This change
introduces two new templates, one for feature requests and one for bug
reports.

It also sets the configuration for the issue picker to send the user to
the appropriate pages to file issues for PowerShell itself, seek support,
or file a documentation issue.