Add support for errors generated by native applications.
This uses redirection from stderr to stdout and then the input which is an error record is added to a new variable __CrescendoErrorQueue.
Two new functions are provided Push-CrescendoError and Pop-CrescendoError.
Push-CrescendoError is used automatically, but Pop-CrescendoError may be used in an output handler to display errors as ErrorRecords with the -EmitAsError parameter.
The default output handler will now automatically write errors using this.
Updated tests to account for these changes where needed.