Add support for argument value transformation. (#175)
* Add support for argument value transformation.

This required a schema change - latest schema is 2022-06
.
We now have 2 new members of the Parameter class `ArgumentTransform` and `ArgumentTransformType` which will invoke a script block which takes the parameter value as an argument.
With this you can now turn `@{ one = 1; two = 2}` into one=1,two=2.
Three different modes are supported. `Inline`, `Function`, and `Script` (similar to output handlers).
Also added changes to the tests to support this new behavior.
Added a new EchoTool program to do native executions.