Add 'ExcludeAsArgument property to parameter.
This allows a parameter to _not_ be passed to the native app.
The variable is still available in the proxy which means it can be used in the output handler.
Fix tests affected by the new property.
Fix mishandling of some parameter transforms.
Fix some descriptions in the schema.
Move ErrorQueue into the proxy rather than a module global, which should provide better isolation.